backend/requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.83
argparse==1.2.1
beautifulsoup4==4.3.2
eventlet==0.15.2
googlemaps==1.0.2
greenlet==0.4.4
gunicorn==19.0.0
httplib2==0.9
itsdangerous==0.24
localtunnel==0.6.1
mongoengine==0.8.7
newrelic==2.32.0.28
nose==1.3.4
oauth2==1.5.211
pymongo==2.7.2
python-gmaps==0.1.1
python-google-places==0.12.1
python-yelp-v2==0.2
pytz==2014.7
requests==2.3.0
six==1.8.0
twilio==3.6.7
unittest2==0.5.1
urllib3==1.7.1
wikipedia==1.3.1
wsgiref==0.1.2
yunomi==0.2.2
